Reorganize fixtures to reduce flakiness
