+++ tests/infrastructure/test_docker_container.py
@@ -6,9 +6,7 @@ from unittest.mock import MagicMock
 
 import anyio.abc
 import docker
-import httpx
 import pytest
-import respx
 
 from prefect.exceptions import InfrastructureNotAvailable, InfrastructureNotFound
 from prefect.infrastructure.container import (
@@ -19,6 +17,9 @@ from prefect.infrastructure.container import (
 )
 from prefect.testing.utilities import assert_does_not_warn
 from prefect.utilities.dockerutils import get_prefect_image_name
+
+# make module level fixtures available
+from .fixtures import *  # noqa
 
 if TYPE_CHECKING:
     from docker import DockerClient
@@ -881,266 +882,91 @@ def test_stream_container_logs_on_real_container(capsys):
     assert "hello" in captured.out
 
 
-default_base_job_template = {
-    "job_configuration": {
-        "command": "{{ command }}",
-        "env": "{{ env }}",
-        "labels": "{{ labels }}",
-        "name": "{{ name }}",
-        "image": "{{ image }}",
-        "image_pull_policy": "{{ image_pull_policy }}",
-        "networks": "{{ networks }}",
-        "network_mode": "{{ network_mode }}",
-        "auto_remove": "{{ auto_remove }}",
-        "volumes": "{{ volumes }}",
-        "stream_output": "{{ stream_output }}",
-        "mem_limit": "{{ mem_limit }}",
-        "memswap_limit": "{{ memswap_limit }}",
-        "privileged": "{{ privileged }}",
-    },
-    "variables": {
-        "description": (
-            "Configuration class used by the Docker worker.\n\nAn instance of this"
-            " class is passed to the Docker worker's `run` method\nfor each flow run."
-            " It contains all the information necessary to execute the\nflow run as a"
-            " Docker container.\n\nAttributes:\n    name: The name to give to created"
-            " Docker containers.\n    command: The command executed in created Docker"
-            " containers to kick off\n        flow run execution.\n    env: The"
-            " environment variables to set in created Docker containers.\n    labels:"
-            " The labels to set on created Docker containers.\n    image: The image"
-            " reference of a container image to use for created jobs.\n        If not"
-            " set, the latest Prefect image will be used.\n    image_pull_policy: The"
-            " image pull policy to use when pulling images.\n    networks: Docker"
-            " networks that created containers should be connected to.\n   "
-            " network_mode: The network mode for the created containers (e.g. host,"
-            " bridge).\n        If 'networks' is set, this cannot be set.\n   "
-            " auto_remove: If set, containers will be deleted on completion.\n   "
-            " volumes: Docker volumes that should be mounted in created containers.\n  "
-            "  stream_output: If set, the output from created containers will be"
-            " streamed\n        to local standard output.\n    mem_limit: Memory limit"
-            " of created containers. Accepts a value\n        with a unit identifier"
-            " (e.g. 100000b, 1000k, 128m, 1g.) If a value is\n        given without a"
-            " unit, bytes are assumed.\n    memswap_limit: Total memory (memory +"
-            " swap), -1 to disable swap. Should only be\n        set if `mem_limit` is"
-            " also set. If `mem_limit` is set, this defaults to\n        allowing the"
-            " container to use as much swap as memory. For example, if\n       "
-            " `mem_limit` is 300m and `memswap_limit` is not set, containers can use\n "
-            "       600m in total of memory and swap.\n    privileged: Give extended"
-            " privileges to created containers."
-        ),
-        "type": "object",
-        "properties": {
-            "command": {
-                "title": "Command",
-                "description": (
-                    "The command to use when starting a flow run. In most cases, this"
-                    " should be left blank and the command will be automatically"
-                    " generated by the worker."
-                ),
-                "type": "string",
-            },
-            "env": {
-                "title": "Environment Variables",
-                "description": "Environment variables to set when starting a flow run.",
-                "type": "object",
-                "additionalProperties": {"type": "string"},
-            },
-            "labels": {
-                "title": "Labels",
-                "description": (
-                    "Labels applied to infrastructure created by the worker using this"
-                    " job configuration."
-                ),
-                "type": "object",
-                "additionalProperties": {"type": "string"},
-            },
-            "name": {
-                "title": "Name",
-                "description": (
-                    "Name given to infrastructure created by the worker using this job"
-                    " configuration."
-                ),
-                "type": "string",
-            },
-            "image": {
-                "title": "Image",
-                "description": (
-                    "The image reference of a container image to use for created jobs."
-                    " If not set, the latest Prefect image will be used."
-                ),
-                "example": "docker.io/prefecthq/prefect:2-latest",
-                "type": "string",
-            },
-            "image_pull_policy": {
-                "title": "Image Pull Policy",
-                "description": "The image pull policy to use when pulling images.",
-                "enum": ["IfNotPresent", "Always", "Never"],
-                "type": "string",
-            },
-            "networks": {
-                "title": "Networks",
-                "description": (
-                    "Docker networks that created containers should be connected to."
-                ),
-                "type": "array",
-                "items": {"type": "string"},
-            },
-            "network_mode": {
-                "title": "Network Mode",
-                "description": (
-                    "The network mode for the created containers (e.g. host, bridge)."
-                    " If 'networks' is set, this cannot be set."
-                ),
-                "type": "string",
-            },
-            "auto_remove": {
-                "title": "Auto Remove",
-                "description": "If set, containers will be deleted on completion.",
-                "default": False,
-                "type": "boolean",
-            },
-            "volumes": {
-                "title": "Volumes",
-                "description": "A list of volume to mount into created containers.",
-                "example": ["/my/local/path:/path/in/container"],
-                "type": "array",
-                "items": {"type": "string"},
-            },
-            "stream_output": {
-                "title": "Stream Output",
-                "description": (
-                    "If set, the output from created containers will be streamed to"
-                    " local standard output."
-                ),
-                "default": True,
-                "type": "boolean",
-            },
-            "mem_limit": {
-                "title": "Memory Limit",
-                "description": (
-                    "Memory limit of created containers. Accepts a value with a unit"
-                    " identifier (e.g. 100000b, 1000k, 128m, 1g.) If a value is given"
-                    " without a unit, bytes are assumed."
-                ),
-                "type": "string",
-            },
-            "memswap_limit": {
-                "title": "Memory Swap Limit",
-                "description": (
-                    "Total memory (memory + swap), -1 to disable swap. Should only be"
-                    " set if `mem_limit` is also set. If `mem_limit` is set, this"
-                    " defaults toallowing the container to use as much swap as memory."
-                    " For example, if `mem_limit` is 300m and `memswap_limit` is not"
-                    " set, containers can use 600m in total of memory and swap."
-                ),
-                "type": "string",
-            },
-            "privileged": {
-                "title": "Privileged",
-                "description": "Give extended privileges to created container.",
-                "default": False,
-                "type": "boolean",
-            },
-        },
-    },
-}
-
-base_job_template_with_defaults = deepcopy(default_base_job_template)
-base_job_template_with_defaults["variables"]["properties"]["command"][
-    "default"
-] = "python my_script.py"
-base_job_template_with_defaults["variables"]["properties"]["env"]["default"] = {
-    "VAR1": "value1",
-    "VAR2": "value2",
-}
-base_job_template_with_defaults["variables"]["properties"]["labels"]["default"] = {
-    "label1": "value1",
-    "label2": "value2",
-}
-base_job_template_with_defaults["variables"]["properties"]["name"][
-    "default"
-] = "my_container"
-base_job_template_with_defaults["variables"]["properties"]["image"][
-    "default"
-] = "docker.io/my_image:latest"
-base_job_template_with_defaults["variables"]["properties"]["image_pull_policy"][
-    "default"
-] = "IfNotPresent"
-base_job_template_with_defaults["variables"]["properties"]["networks"]["default"] = [
-    "network1",
-    "network2",
-]
-base_job_template_with_defaults["variables"]["properties"]["network_mode"][
-    "default"
-] = "bridge"
-base_job_template_with_defaults["variables"]["properties"]["auto_remove"][
-    "default"
-] = True
-base_job_template_with_defaults["variables"]["properties"]["volumes"]["default"] = [
-    "/path1:/container/path1",
-    "/path2:/container/path2",
-]
-base_job_template_with_defaults["variables"]["properties"]["stream_output"][
-    "default"
-] = True
-base_job_template_with_defaults["variables"]["properties"]["mem_limit"][
-    "default"
-] = "1g"
-base_job_template_with_defaults["variables"]["properties"]["memswap_limit"][
-    "default"
-] = "2g"
-base_job_template_with_defaults["variables"]["properties"]["privileged"][
-    "default"
-] = False
-
-
-@pytest.fixture()
-async def mock_collection_registry():
-    mock_body = {
-        "prefect-docker": {
-            "docker": {
-                "type": "docker",
-                "default_base_job_configuration": default_base_job_template,
-            }
-        },
+@pytest.fixture
+def base_job_template_with_defaults(docker_default_base_job_template):
+    base_job_template_with_defaults = deepcopy(docker_default_base_job_template)
+    base_job_template_with_defaults["variables"]["properties"]["command"][
+        "default"
+    ] = "python my_script.py"
+    base_job_template_with_defaults["variables"]["properties"]["env"]["default"] = {
+        "VAR1": "value1",
+        "VAR2": "value2",
     }
-    with respx.mock(
-        assert_all_mocked=True,
-        assert_all_called=False,
-    ) as respx_mock:
-        respx_mock.get(
-            "https://raw.githubusercontent.com/PrefectHQ/prefect-collection-registry/main/views/aggregate-worker-metadata.json"
-        ).mock(return_value=httpx.Response(200, json=mock_body))
-        yield
+    base_job_template_with_defaults["variables"]["properties"]["labels"]["default"] = {
+        "label1": "value1",
+        "label2": "value2",
+    }
+    base_job_template_with_defaults["variables"]["properties"]["name"][
+        "default"
+    ] = "my_container"
+    base_job_template_with_defaults["variables"]["properties"]["image"][
+        "default"
+    ] = "docker.io/my_image:latest"
+    base_job_template_with_defaults["variables"]["properties"]["image_pull_policy"][
+        "default"
+    ] = "IfNotPresent"
+    base_job_template_with_defaults["variables"]["properties"]["networks"][
+        "default"
+    ] = [
+        "network1",
+        "network2",
+    ]
+    base_job_template_with_defaults["variables"]["properties"]["network_mode"][
+        "default"
+    ] = "bridge"
+    base_job_template_with_defaults["variables"]["properties"]["auto_remove"][
+        "default"
+    ] = True
+    base_job_template_with_defaults["variables"]["properties"]["volumes"]["default"] = [
+        "/path1:/container/path1",
+        "/path2:/container/path2",
+    ]
+    base_job_template_with_defaults["variables"]["properties"]["stream_output"][
+        "default"
+    ] = True
+    base_job_template_with_defaults["variables"]["properties"]["mem_limit"][
+        "default"
+    ] = "1g"
+    base_job_template_with_defaults["variables"]["properties"]["memswap_limit"][
+        "default"
+    ] = "2g"
+    base_job_template_with_defaults["variables"]["properties"]["privileged"][
+        "default"
+    ] = False
+    return base_job_template_with_defaults
 
 
 @pytest.mark.usefixtures("mock_collection_registry")
 @pytest.mark.parametrize(
-    "container,expected_template",
+    "container_config",
     [
-        (DockerContainer(), default_base_job_template),
-        (
-            DockerContainer(
-                command=["python", "my_script.py"],
-                env={"VAR1": "value1", "VAR2": "value2"},
-                labels={"label1": "value1", "label2": "value2"},
-                name="my_container",
-                image="docker.io/my_image:latest",
-                image_pull_policy="IF_NOT_PRESENT",
-                networks=["network1", "network2"],
-                network_mode="bridge",
-                auto_remove=True,
-                volumes=["/path1:/container/path1", "/path2:/container/path2"],
-                stream_output=True,
-                mem_limit="1g",
-                memswap_limit="2g",
-                privileged=False,
-            ),
-            base_job_template_with_defaults,
-        ),
+        "default",
+        "custom",
     ],
 )
-async def test_generate_work_pool_base_job_template(container, expected_template):
+async def test_generate_work_pool_base_job_template(
+    container_config, docker_default_base_job_template, base_job_template_with_defaults
+):
+    container = DockerContainer()
+    expected_template = docker_default_base_job_template
+    if container_config == "custom":
+        container = DockerContainer(
+            command=["python", "my_script.py"],
+            env={"VAR1": "value1", "VAR2": "value2"},
+            labels={"label1": "value1", "label2": "value2"},
+            name="my_container",
+            image="docker.io/my_image:latest",
+            image_pull_policy="IF_NOT_PRESENT",
+            networks=["network1", "network2"],
+            network_mode="bridge",
+            auto_remove=True,
+            volumes=["/path1:/container/path1", "/path2:/container/path2"],
+            stream_output=True,
+            mem_limit="1g",
+            memswap_limit="2g",
+            privileged=False,
+        )
+        expected_template = base_job_template_with_defaults
     template = await container.generate_work_pool_base_job_template()
 
     assert template == expected_template
